# Optimize top_set_bit(::BigInt) and hash(::Real)