EDIT: this PR now does a fair bit more than #49986 did, see later comments.

Re-lands #49986, but computes `abs` first. This should be safe now because `ndigits0z(x, 2)` and `top_set_bit(abs(x))` are semantically equivalent whereas before I was falsely assuming `x > 0`.

The performance gain is now only 1.4x instead of 1.5x, which could be noise or the cost of `abs` (probably noise)