@LilithHafner 2023-05-30
Empirical check that `top_set_bit(abs(x)) == Base.ndigits0z(x, 2)` is valid:

```julia
for T in [UInt8, Int8] 
    for i in typemin(T):typemax(T)
        for j in (i, big(i), Int(i), unsigned(Int(i)))
            Base.top_set_bit(abs(j)) == Base.ndigits0z(j, 2) || error()
        end
    end
end
```