@oscardssmith 2023-05-30
While we're at it, is there a reason we're doing `hash(Int64(num) << Int(pow), h)` rather than `hash(Int64(x))`? (and similarly `hash(Float64(x))` instead of `hash(Int64(num) << Int(pow), h)`