@oscardssmith 2023-05-30
interesting. That somehow seems to be faster than https://github.com/JuliaLang/julia/pull/49998 which is rather surprising to me. I guess our compiler is doing a really good job cleaning this up for `Int128`.