@oscardssmith 2023-06-02
Closing since https://github.com/JuliaLang/julia/pull/49996 gets the same performance.